refactor(extension): reduce execution.rs item count under threshold (#6276)
